split OrganizationCommander -> OrgCommander, OrgInfoCommander
