+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/OrganizationCommander.scala
@@ -1,8 +1,7 @@
 package com.keepit.commanders
 
-import com.google.inject.{ ImplementedBy, Inject, Singleton }
+import com.google.inject.{ ImplementedBy, Inject }
 import com.keepit.common.controller.UserRequest
-import com.keepit.common.crypto.PublicIdConfiguration
 import com.keepit.common.db.Id
 import com.keepit.common.db.slick.DBSession.{ RSession, RWSession }
 import com.keepit.common.db.slick.Database
@@ -10,239 +9,48 @@ import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.common.logging.Logging
 import com.keepit.common.net.URI
 import com.keepit.common.performance.{ StatsdTiming, AlertingTimer }
-import com.keepit.common.social.BasicUserRepo
-import com.keepit.common.store.ImageSize
 import com.keepit.eliza.ElizaServiceClient
 import com.keepit.heimdal.HeimdalContext
 import com.keepit.model.OrganizationPermission.{ MANAGE_PLAN, EDIT_ORGANIZATION }
 import com.keepit.model._
-import com.keepit.social.BasicUser
-import com.keepit.payments._
+import com.keepit.payments.{ CreditRewardCommander, RewardTrigger, ActionAttribution, PaidPlan, PaidPlanRepo, PlanManagementCommander }
 import play.api.Play
+import scala.concurrent.duration._
 
 import scala.concurrent.{ ExecutionContext, Future }
 
 @ImplementedBy(classOf[OrganizationCommanderImpl])
 trait OrganizationCommander {
-  def getOrganizationView(orgId: Id[Organization], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String]): OrganizationView
-  def getOrganizationViews(orgIds: Set[Id[Organization]], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String]): Map[Id[Organization], OrganizationView]
-  def getOrganizationViewHelper(orgId: Id[Organization], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String])(implicit session: RSession): OrganizationView
-  def getBasicOrganizationView(orgId: Id[Organization], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String]): BasicOrganizationView
-  def getBasicOrganizationViewsHelper(orgIds: Set[Id[Organization]], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String])(implicit session: RSession): Map[Id[Organization], BasicOrganizationView]
-  def getBasicOrganizationViewHelper(orgId: Id[Organization], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String])(implicit session: RSession): BasicOrganizationView
-  def getOrganizationInfo(orgId: Id[Organization], viewerIdOpt: Option[Id[User]])(implicit session: RSession): OrganizationInfo
-  def getOrganizationInfos(orgIds: Set[Id[Organization]], viewerIdOpt: Option[Id[User]]): Map[Id[Organization], OrganizationInfo]
-  def getAccountFeatureSettings(orgId: Id[Organization]): OrganizationSettingsResponse
-  def getExternalOrgConfiguration(orgId: Id[Organization]): ExternalOrganizationConfiguration
-  def getExternalOrgConfigurationHelper(orgId: Id[Organization])(implicit session: RSession): ExternalOrganizationConfiguration
-  def getBasicOrganizationHelper(orgId: Id[Organization])(implicit session: RSession): Option[BasicOrganization]
-  def getBasicOrganizations(orgIds: Set[Id[Organization]]): Map[Id[Organization], BasicOrganization]
-  def getOrganizationLibrariesVisibleToUser(orgId: Id[Organization], userIdOpt: Option[Id[User]], offset: Offset, limit: Limit): Seq[LibraryCardInfo]
-  def getLibrariesVisibleToUserHelper(orgId: Id[Organization], userIdOpt: Option[Id[User]], offset: Offset, limit: Limit)(implicit session: RSession): Seq[Library]
   def createOrganization(request: OrganizationCreateRequest)(implicit eventContext: HeimdalContext): Either[OrganizationFail, OrganizationCreateResponse]
   def modifyOrganization(request: OrganizationModifyRequest)(implicit eventContext: HeimdalContext): Either[OrganizationFail, OrganizationModifyResponse]
-  def setAccountFeatureSettings(request: OrganizationSettingsRequest): Either[OrganizationFail, OrganizationSettingsResponse]
-  def deleteOrganization(request: OrganizationDeleteRequest)(implicit eventContext: HeimdalContext): Either[OrganizationFail, OrganizationDeleteResponse]
   def transferOrganization(request: OrganizationTransferRequest)(implicit eventContext: HeimdalContext): Either[OrganizationFail, OrganizationTransferResponse]
-
-  def unsafeModifyOrganization(request: UserRequest[_], orgId: Id[Organization], modifications: OrganizationModifications): Unit
+  def setAccountFeatureSettings(req: OrganizationSettingsRequest): Either[OrganizationFail, OrganizationSettingsResponse]
   def unsafeSetAccountFeatureSettings(orgId: Id[Organization], settings: OrganizationSettings)(implicit session: RWSession): OrganizationSettingsResponse
-  def getOrgTrackingValues(orgId: Id[Organization]): OrgTrackingValues
-
+  def deleteOrganization(request: OrganizationDeleteRequest)(implicit eventContext: HeimdalContext): Either[OrganizationFail, OrganizationDeleteResponse]
+  def unsafeModifyOrganization(request: UserRequest[_], orgId: Id[Organization], modifications: OrganizationModifications): Unit
 }
 
-@Singleton
 class OrganizationCommanderImpl @Inject() (
     db: Database,
-    permissionCommander: PermissionCommander,
     orgRepo: OrganizationRepo,
-    orgConfigRepo: OrganizationConfigurationRepo,
-    paidPlanRepo: PaidPlanRepo,
-    orgMembershipRepo: OrganizationMembershipRepo,
-    orgMembershipCommander: OrganizationMembershipCommander,
-    orgInviteCommander: OrganizationInviteCommander,
     organizationMembershipCandidateRepo: OrganizationMembershipCandidateRepo,
-    organizationDomainOwnershipCommander: OrganizationDomainOwnershipCommander,
+    orgMembershipRepo: OrganizationMembershipRepo,
     orgInviteRepo: OrganizationInviteRepo,
-    userExperimentRepo: UserExperimentRepo,
-    organizationAvatarCommander: OrganizationAvatarCommander,
+    orgConfigRepo: OrganizationConfigurationRepo,
     userRepo: UserRepo,
-    keepRepo: KeepRepo,
+    paidPlanRepo: PaidPlanRepo,
     libraryRepo: LibraryRepo,
-    basicUserRepo: BasicUserRepo,
-    creditRewardCommander: CreditRewardCommander,
-    libraryMembershipRepo: LibraryMembershipRepo,
-    libraryCardCommander: LibraryCardCommander,
-    libraryCommander: LibraryCommander,
-    airbrake: AirbrakeNotifier,
-    orgExperimentRepo: OrganizationExperimentRepo,
-    organizationAnalytics: OrganizationAnalytics,
-    implicit val publicIdConfig: PublicIdConfiguration,
-    handleCommander: HandleCommander,
+    organizationInfoCommander: OrganizationInfoCommander,
     planManagementCommander: PlanManagementCommander,
-    basicOrganizationIdCache: BasicOrganizationIdCache,
+    permissionCommander: PermissionCommander,
+    handleCommander: HandleCommander,
+    libraryCommander: LibraryCommander,
+    orgMembershipCommander: OrganizationMembershipCommander,
+    creditRewardCommander: CreditRewardCommander,
+    organizationAnalytics: OrganizationAnalytics,
     eliza: ElizaServiceClient,
+    airbrake: AirbrakeNotifier,
     implicit val executionContext: ExecutionContext) extends OrganizationCommander with Logging {
-
-  def getOrganizationView(orgId: Id[Organization], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String]): OrganizationView = {
-    db.readOnlyReplica { implicit session => getOrganizationViewHelper(orgId, viewerIdOpt, authTokenOpt) }
-  }
-
-  def getOrganizationViews(orgIds: Set[Id[Organization]], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String]): Map[Id[Organization], OrganizationView] = {
-    db.readOnlyReplica { implicit session => orgIds.map(id => id -> getOrganizationViewHelper(id, viewerIdOpt, authTokenOpt)).toMap }
-  }
-
-  def getOrganizationViewHelper(orgId: Id[Organization], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String])(implicit session: RSession): OrganizationView = {
-    val organizationInfo = getOrganizationInfo(orgId, viewerIdOpt)
-    val membershipInfo = getMembershipInfoHelper(orgId, viewerIdOpt, authTokenOpt)
-    OrganizationView(organizationInfo, membershipInfo)
-  }
-
-  def getBasicOrganizationView(orgId: Id[Organization], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String]): BasicOrganizationView = {
-    db.readOnlyReplica { implicit session => getBasicOrganizationViewHelper(orgId, viewerIdOpt, authTokenOpt) }
-  }
-
-  def getBasicOrganizationViewsHelper(orgIds: Set[Id[Organization]], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String])(implicit session: RSession): Map[Id[Organization], BasicOrganizationView] = {
-    orgIds.map(id => id -> getBasicOrganizationViewHelper(id, viewerIdOpt, authTokenOpt)).toMap
-  }
-
-  def getBasicOrganizationViewHelper(orgId: Id[Organization], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String])(implicit session: RSession): BasicOrganizationView = {
-    // This function assumes that the org is active
-    val basicOrganization = basicOrganizationIdCache.getOrElse(BasicOrganizationIdKey(orgId))(getBasicOrganizationHelper(orgId).get)
-    val membershipInfo = getMembershipInfoHelper(orgId, viewerIdOpt, authTokenOpt)
-    BasicOrganizationView(basicOrganization, membershipInfo)
-  }
-
-  def getBasicOrganizations(orgIds: Set[Id[Organization]]): Map[Id[Organization], BasicOrganization] = {
-    val cacheFormattedMap = db.readOnlyReplica { implicit session =>
-      basicOrganizationIdCache.bulkGetOrElse(orgIds.map(BasicOrganizationIdKey)) { missing =>
-        missing.map(_.id).map {
-          orgId => orgId -> getBasicOrganizationHelper(orgId) // grab all the Option[BasicOrganization]
-        }.collect {
-          case (orgId, Some(basicOrg)) => orgId -> basicOrg // take only the active orgs (inactive ones are None)
-        }.map {
-          case (orgId, org) => (BasicOrganizationIdKey(orgId), org) // format them so the cache can understand them
-        }.toMap
-      }
-    }
-    cacheFormattedMap.map { case (orgKey, org) => (orgKey.id, org) }
-  }
-
-  def getBasicOrganizationHelper(orgId: Id[Organization])(implicit session: RSession): Option[BasicOrganization] = {
-    val org = orgRepo.get(orgId)
-    if (org.isInactive) None
-    else {
-      val orgHandle = org.handle
-      val orgName = org.name
-      val description = org.description
-
-      val ownerId = userRepo.get(org.ownerId).externalId
-      val avatarPath = organizationAvatarCommander.getBestImageByOrgId(orgId, ImageSize(200, 200)).imagePath
-
-      Some(BasicOrganization(
-        orgId = Organization.publicId(orgId),
-        ownerId = ownerId,
-        handle = orgHandle,
-        name = orgName,
-        description = description,
-        avatarPath = avatarPath))
-    }
-  }
-
-  def getOrganizationInfos(orgIds: Set[Id[Organization]], viewerIdOpt: Option[Id[User]]): Map[Id[Organization], OrganizationInfo] = {
-    db.readOnlyReplica { implicit session =>
-      orgIds.map(id => id -> getOrganizationInfo(id, viewerIdOpt)).toMap
-    }
-  }
-
-  def getAccountFeatureSettings(orgId: Id[Organization]): OrganizationSettingsResponse = {
-    db.readOnlyReplica { implicit session =>
-      val config = orgConfigRepo.getByOrgId(orgId)
-      OrganizationSettingsResponse(config)
-    }
-  }
-
-  def getExternalOrgConfiguration(orgId: Id[Organization]): ExternalOrganizationConfiguration = {
-    db.readOnlyReplica(implicit session => getExternalOrgConfigurationHelper(orgId))
-  }
-
-  def getExternalOrgConfigurationHelper(orgId: Id[Organization])(implicit session: RSession): ExternalOrganizationConfiguration = {
-    val config = orgConfigRepo.getByOrgId(orgId)
-    val plan = planManagementCommander.currentPlanHelper(orgId)
-    ExternalOrganizationConfiguration(plan.showUpsells, OrganizationSettingsWithEditability(config.settings, plan.editableFeatures))
-  }
-
-  def getOrganizationInfo(orgId: Id[Organization], viewerIdOpt: Option[Id[User]])(implicit session: RSession): OrganizationInfo = {
-    val viewerPermissions = permissionCommander.getOrganizationPermissions(orgId, viewerIdOpt)
-    if (!viewerPermissions.contains(OrganizationPermission.VIEW_ORGANIZATION)) {
-      airbrake.notify(s"Tried to serve up organization info for org $orgId to viewer $viewerIdOpt, but they do not have permission to view this org")
-    }
-
-    val org = orgRepo.get(orgId)
-    if (org.state == OrganizationStates.INACTIVE) throw new Exception(s"inactive org: $org")
-    val orgHandle = org.handle
-    val orgName = org.name
-    val description = org.description
-    val site = org.site
-    val ownerId = userRepo.get(org.ownerId).externalId
-
-    val memberIds = {
-      if (!viewerPermissions.contains(OrganizationPermission.VIEW_MEMBERS)) Seq.empty
-      else orgMembershipRepo.getSortedMembershipsByOrgId(orgId, Offset(0), Limit(Int.MaxValue)).map(_.userId)
-    }
-    val members = userRepo.getAllUsers(memberIds).values.toSeq
-    val membersAsBasicUsers = members.map(BasicUser.fromUser)
-    val memberCount = members.length
-    val avatarPath = organizationAvatarCommander.getBestImageByOrgId(orgId, ImageSize(200, 200)).imagePath
-    val config = getExternalOrgConfigurationHelper(orgId)
-    val numLibraries = countLibrariesVisibleToUserHelper(orgId, viewerIdOpt)
-
-    OrganizationInfo(
-      orgId = Organization.publicId(orgId),
-      ownerId = ownerId,
-      handle = orgHandle,
-      name = orgName,
-      description = description,
-      site = site,
-      avatarPath = avatarPath,
-      members = membersAsBasicUsers,
-      numMembers = memberCount,
-      numLibraries = numLibraries,
-      config = config)
-  }
-
-  private def getMembershipInfoHelper(orgId: Id[Organization], viewerIdOpt: Option[Id[User]], authTokenOpt: Option[String])(implicit session: RSession): OrganizationViewerInfo = {
-    val membershipOpt = viewerIdOpt.flatMap(orgMembershipRepo.getByOrgIdAndUserId(orgId, _))
-    val inviteOpt = orgInviteCommander.getViewerInviteInfo(orgId, viewerIdOpt, authTokenOpt)
-    val sharedEmails = viewerIdOpt.map(userId => organizationDomainOwnershipCommander.getSharedEmailsHelper(userId, orgId)).getOrElse(Set.empty)
-    val permissions = permissionCommander.getOrganizationPermissions(orgId, viewerIdOpt)
-
-    OrganizationViewerInfo(
-      invite = inviteOpt,
-      emails = sharedEmails,
-      permissions = permissions,
-      membership = membershipOpt.map(mem => OrganizationMembershipInfo(mem.role))
-    )
-  }
-
-  def getOrganizationLibrariesVisibleToUser(orgId: Id[Organization], userIdOpt: Option[Id[User]], offset: Offset, limit: Limit): Seq[LibraryCardInfo] = {
-    db.readOnlyReplica { implicit session =>
-      val visibleLibraries = getLibrariesVisibleToUserHelper(orgId, userIdOpt, offset, limit)
-      val basicOwnersByOwnerId = basicUserRepo.loadAll(visibleLibraries.map(_.ownerId).toSet)
-      libraryCardCommander.createLibraryCardInfos(visibleLibraries, basicOwnersByOwnerId, userIdOpt, withFollowing = false, ProcessedImageSize.Medium.idealSize).seq
-    }
-  }
-
-  def getLibrariesVisibleToUserHelper(orgId: Id[Organization], userIdOpt: Option[Id[User]], offset: Offset, limit: Limit)(implicit session: RSession): Seq[Library] = {
-    val viewerLibraryMemberships = userIdOpt.map(libraryMembershipRepo.getWithUserId(_).map(_.libraryId).toSet).getOrElse(Set.empty[Id[Library]])
-    val includeOrgVisibleLibs = userIdOpt.exists(orgMembershipRepo.getByOrgIdAndUserId(orgId, _).isDefined)
-    libraryRepo.getVisibleOrganizationLibraries(orgId, includeOrgVisibleLibs, viewerLibraryMemberships, offset, limit)
-  }
-  private def countLibrariesVisibleToUserHelper(orgId: Id[Organization], userIdOpt: Option[Id[User]])(implicit session: RSession): Int = {
-    val viewerLibraryMemberships = userIdOpt.map(libraryMembershipRepo.getWithUserId(_).map(_.libraryId).toSet).getOrElse(Set.empty[Id[Library]])
-    val includeOrgVisibleLibs = userIdOpt.exists(orgMembershipRepo.getByOrgIdAndUserId(orgId, _).isDefined)
-    libraryRepo.countVisibleOrganizationLibraries(orgId, includeOrgVisibleLibs, viewerLibraryMemberships)
-  }
 
   private def getValidationError(request: OrganizationRequest)(implicit session: RSession): Option[OrganizationFail] = {
     request match {
@@ -325,7 +133,7 @@ class OrganizationCommanderImpl @Inject() (
             eliza.flush(request.requesterId)
           }
 
-          val orgView = getOrganizationViewHelper(org.id.get, Some(request.requesterId), None)
+          val orgView = organizationInfoCommander.getOrganizationViewHelper(org.id.get, Some(request.requesterId), None)
           Right(OrganizationCreateResponse(request, org, orgGeneralLibrary, orgView))
       }
     }
@@ -341,12 +149,13 @@ class OrganizationCommanderImpl @Inject() (
             creditRewardCommander.registerRewardTrigger(RewardTrigger.OrganizationDescriptionAdded(org.id.get, modifiedOrg))
           }
           organizationAnalytics.trackOrganizationEvent(org, userRepo.get(request.requesterId), request)
-          val orgView = getOrganizationViewHelper(request.orgId, Some(request.requesterId), None)
+          val orgView = organizationInfoCommander.getOrganizationViewHelper(request.orgId, Some(request.requesterId), None)
           Right(OrganizationModifyResponse(request, orgRepo.save(modifiedOrg), orgView))
         case Some(orgFail) => Left(orgFail)
       }
     }
   }
+
   def setAccountFeatureSettings(req: OrganizationSettingsRequest): Either[OrganizationFail, OrganizationSettingsResponse] = {
     db.readWrite { implicit session =>
       getValidationError(req) match {
@@ -444,17 +253,6 @@ class OrganizationCommanderImpl @Inject() (
     db.readWrite { implicit session =>
       val org = orgRepo.get(orgId)
       val modifiedOrg = orgRepo.save(organizationWithModifications(org, modifications))
-    }
-  }
-
-  def getOrgTrackingValues(orgId: Id[Organization]): OrgTrackingValues = {
-    db.readOnlyReplica { implicit session =>
-      val libraries = libraryRepo.getOrganizationLibraries(orgId)
-      val libraryCount = libraries.length
-      val keepCount = keepRepo.getByLibraryIds(libraries.map(_.id.get).toSet).count(keep => !KeepSource.imports.contains(keep.source))
-      val inviteCount = orgInviteRepo.getCountByOrganization(orgId, decisions = Set(InvitationDecision.PENDING))
-      val collabLibCount = libraryMembershipRepo.countWithAccessByLibraryId(libraries.map(_.id.get).toSet, LibraryAccess.READ_WRITE).count { case (_, memberCount) => memberCount > 0 }
-      OrgTrackingValues(libraryCount, keepCount, inviteCount, collabLibCount)
     }
   }
 }
