+++ kifi-backend/modules/shoebox/app/com/keepit/common/social/SocialGraphPluginImpl.scala
@@ -42,7 +42,6 @@ private[social] class SocialGraphActor @Inject() (
   socialUserCreateConnections: UserConnectionCreator,
   socialUserTypeahead: SocialUserTypeahead,
   userValueRepo: UserValueRepo,
-  userCommander: UserCommander,
   emailAddressCommander: UserEmailAddressCommander,
   heimdal: HeimdalServiceClient,
   clock: Clock)
