+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/ext/ExtAuthController.scala
@@ -32,7 +32,7 @@ class ExtAuthController @Inject() (
   authCommander: AuthCommander,
   libraryInfoCommander: LibraryInfoCommander,
   installationRepo: KifiInstallationRepo,
-  orgCommander: OrganizationCommander,
+  orgInfoCommander: OrganizationInfoCommander,
   orgMembershipCommander: OrganizationMembershipCommander,
   urlPatternRepo: URLPatternRepo,
   experimentCommander: LocalUserExperimentCommander,
@@ -77,7 +77,7 @@ class ExtAuthController @Inject() (
         val (libraries, organizations, installation, urlPatterns, isInstall, isUpdate) = db.readWrite { implicit s =>
           val libraries = libraryInfoCommander.getMainAndSecretLibrariesForUser(userId)
           val orgIds = orgMembershipCommander.getAllOrganizationsForUser(userId)
-          val basicOrgs = orgCommander.getBasicOrganizationViewsHelper(orgIds.toSet, Some(userId), authTokenOpt = None).values.toSeq
+          val basicOrgs = orgInfoCommander.getBasicOrganizationViewsHelper(orgIds.toSet, Some(userId), authTokenOpt = None).values.toSeq
           val (installation, isInstall, isUpdate): (KifiInstallation, Boolean, Boolean) = installationIdOpt flatMap { id =>
             installationRepo.getOpt(userId, id)
           } match {
