+++ kifi-backend/modules/shoebox/app/com/keepit/payments/ActivityLogCommander.scala
@@ -1,7 +1,7 @@
 package com.keepit.payments
 
 import com.google.inject.{ Singleton, Inject, ImplementedBy }
-import com.keepit.commanders.OrganizationCommander
+import com.keepit.commanders.{ OrganizationAvatarCommander, OrganizationCommander }
 import com.keepit.common.crypto.{ PublicIdConfiguration, PublicId }
 import com.keepit.common.db.Id
 import com.keepit.common.db.slick.DBSession.RSession
@@ -9,6 +9,7 @@ import com.keepit.common.db.slick.Database
 import com.keepit.common.mail.{ SystemEmailAddress, EmailAddress }
 import com.keepit.common.path.Path
 import com.keepit.common.social.BasicUserRepo
+import com.keepit.common.store.ImageSize
 import com.keepit.model._
 import com.keepit.social.BasicUser
 import org.joda.time.DateTime
@@ -33,7 +34,7 @@ class ActivityLogCommanderImpl @Inject() (
     creditRewardInfoCommander: CreditRewardInfoCommander,
     basicUserRepo: BasicUserRepo,
     organizationRepo: OrganizationRepo,
-    organizationCommander: OrganizationCommander,
+    organizationAvatarCommander: OrganizationAvatarCommander,
     implicit val publicIdConfig: PublicIdConfiguration) extends ActivityLogCommander {
 
   private def orgId2AccountId(orgId: Id[Organization])(implicit session: RSession): Id[PaidAccount] = {
@@ -46,7 +47,20 @@ class ActivityLogCommanderImpl @Inject() (
   }
 
   private def getUser(id: Id[User])(implicit session: RSession): BasicUser = basicUserRepo.load(id)
-  private def getOrg(id: Id[Organization])(implicit session: RSession): BasicOrganization = organizationCommander.getBasicOrganizationHelper(id).getOrElse(throw new Exception(s"Tried to build event info for dead org: $id"))
+  private def getOrg(id: Id[Organization])(implicit session: RSession): BasicOrganization = {
+    val org = organizationRepo.get(id)
+    if (org.isInactive) throw new Exception(s"Tried to build event info for dead org: $id")
+    else {
+      BasicOrganization(
+        orgId = Organization.publicId(org.id.get),
+        ownerId = basicUserRepo.load(org.ownerId).externalId,
+        handle = org.handle,
+        name = org.name,
+        description = org.description,
+        avatarPath = organizationAvatarCommander.getBestImageByOrgId(org.id.get, ImageSize(200, 200)).imagePath
+      )
+    }
+  }
 
   def buildSimpleEventInfo(event: AccountEvent): SimpleAccountEventInfo = db.readOnlyMaster { implicit s =>
     buildSimpleEventInfoHelper(event)
